Add simple initial docker build step

See merge request smart/watch!34

+++ Dockerfile
@@ -1,24 +1,49 @@
 FROM nvidia/cuda:10.2-cudnn8-devel-ubuntu18.04
 
-RUN apt-get update --fix-missing && \
-    apt-get install -y wget bzip2 ca-certificates git libgl1-mesa-glx
+ARG BUILD_STRICT=0
 
-# To update to a newer version see:
-# https://docs.conda.io/en/latest/miniconda_hashes.html for updating
-RUN wget --quiet https://repo.anaconda.com/miniconda/Miniconda3-py38_4.9.2-Linux-x86_64.sh -O ~/miniconda.sh && \
-    /bin/bash ~/miniconda.sh -b -p /opt/conda && \
-    rm ~/miniconda.sh && \
-    /opt/conda/bin/conda clean -tipsy && \
+RUN apt-get update -q && \
+    apt-get install -q -y --no-install-recommends \
+        bzip2 \
+        ca-certificates \
+        git \
+        libgl1-mesa-glx \
+        libglib2.0-0 \
+        libsm6 \
+        libxext6 \
+        libxrender1 \
+        mercurial \
+        subversion \
+        wget \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+ENV PATH /opt/conda/bin:$PATH
+
+CMD [ "/bin/bash" ]
+
+# Leave these args here to better use the Docker build cache
+ARG CONDA_VERSION=py38_4.9.2
+ARG CONDA_MD5=122c8c9beb51e124ab32a0fa6426c656
+
+RUN wget --quiet https://repo.anaconda.com/miniconda/Miniconda3-${CONDA_VERSION}-Linux-x86_64.sh -O miniconda.sh && \
+    echo "${CONDA_MD5}  miniconda.sh" > miniconda.md5 && \
+    if ! md5sum --status -c miniconda.md5; then exit 1; fi && \
+    mkdir -p /opt && \
+    sh miniconda.sh -b -p /opt/conda && \
+    rm miniconda.sh miniconda.md5 && \
     ln -s /opt/conda/etc/profile.d/conda.sh /etc/profile.d/conda.sh && \
     echo ". /opt/conda/etc/profile.d/conda.sh" >> ~/.bashrc && \
     echo "conda activate watch" >> ~/.bashrc && \
-    /opt/conda/bin/conda config --set channel_priority flexible && \
-    /opt/conda/bin/conda update -n base conda -y && \
+    find /opt/conda/ -follow -type f -name '*.a' -delete && \
+    find /opt/conda/ -follow -type f -name '*.js.map' -delete && \
     /opt/conda/bin/conda clean -afy
 
 SHELL ["/bin/bash", "--login", "-c"]
 
 COPY conda_env.yml /watch/
+
+RUN if [ "$BUILD_STRICT" -eq 1 ]; then sed -i 's/>=/==/g' /watch/conda_env.yml; fi
 
 RUN conda env create -f /watch/conda_env.yml
